05.00 need css

+++ eww/scripts/battery.sh
@@ -1,37 +1,46 @@
 #!/bin/sh
 
+#󰁺󰁻󰁼󰁽󰁾󰁿󰂀󰂁󰂂󰁹󰂄
+
 while true; do
-	level=$(cat /sys/class/power_supply/BAT0/capacity)
-	state=$(cat /sys/class/power_supply/BAT0/status)
-	file=/tmp/battery
+    level=$(cat /sys/class/power_supply/BAT0/capacity)
+    state=$(cat /sys/class/power_supply/BAT0/status)
+    file=/tmp/battery
 
-	if [ -f "$file" ]; then
-		if [ "$state" = "Charging" ]; then
-			rm /tmp/battery
-			echo '(battery :color "yellow" :charging true :icon "󰂄" :level '$level')'
+    if [ "$state" = "Charging" ]; then
+        icon="󰂄"
+    elif [ "$level" -gt 90 ]; then
+        icon="󰁹"
+    elif [ "$level" -gt 80 ]; then
+        icon="󰂂"
+    elif [ "$level" -gt 70 ]; then
+        icon="󰂁"
+    elif [ "$level" -gt 60 ]; then
+        icon="󰂀"
+    elif [ "$level" -gt 50 ]; then
+        icon="󰁿"
+    elif [ "$level" -gt 40 ]; then
+        icon="󰁾"
+    elif [ "$level" -gt 30 ]; then
+        icon="󰁽"
+    elif [ "$level" -gt 20 ]; then
+        icon="󰁼"
+    elif [ "$level" -gt 10 ]; then
+        icon="󰁻"
+    else
+        icon="󰁺"
+    fi
 
-		elif [ "$level" -gt 30 ]; then
-			rm /tmp/battery
-			echo '(battery :color "red" :charging false :icon "󰁹" :level '$level')'
+    if [ -f $file ] && [ "$state" = "Charging" ]; then
+        rm /tmp/battery
+    fi
 
-		else
-			echo '(battery :color "green" :charging false :icon "󰁼" :level '$level')'
-		fi
+    if [ "$level" -lt 31 ] && ! [ -f file ] && ! [ "$state" = "Charging" ]; then
+        #notify-send "test"
+        touch /tmp/battery
+    fi
 
-	else
-		if [ "$state" = "Charging" ]; then
-			echo '(battery :color "yellow" :charging true :icon "󰂄" :level '$level')'
+    echo "(battery :icon '${icon}' :level '${level}')"
 
-		elif [ "$level" -gt 30 ]; then
-			touch /tmp/battery
-			#notify-send "tset"
-
-			echo '(battery :color "red" :charging false :icon "󰁹" :level '$level')'
-
-		else
-			echo '(battery :color "green" :charging false :icon "󰁼" :level '$level')'
-		fi
-	fi
-
-	sleep 10
+    sleep 10 
 done
